requirements.txt
langchain
langchain-google-genai
langchain-openai
python-dotenv
flask
pydantic
google-generativeai
